+++ packages/config-ui/src/index.js
@@ -15,9 +15,13 @@ export TwoChoice, { NChoice } from './two-choice';
 export TagsInput from './tags-input';
 export MuiBox from './mui-box';
 export ChoiceConfiguration from './choice-configuration';
-export ConfigLayout from './layout/config-layout';
-export PreviewLayout from './layout/preview-layout';
+export * as layout from './layout';
+
 export * as choiceUtils from './choice-utils';
 export withStatefulModel from './with-stateful-model';
 export Toggle from './settings/toggle';
 export DisplaySize from './settings/display-size';
+
+import * as settings from './settings';
+
+export { settings };