+++ packages/config-ui/src/layout/index.js
@@ -1,0 +1,4 @@
+import ConfigLayout from './config-layout';
+import LayoutContents from './layout-contents';
+
+export { LayoutContents, ConfigLayout };