+++ packages/config-ui/CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.0.1](https://github.com/pie-framework/pie-lib/compare/@pie-lib/config-ui@8.0.0...@pie-lib/config-ui@9.0.1) (2019-04-12)
+
+
+### Bug Fixes
+
+* bump version ([2bc9035](https://github.com/pie-framework/pie-lib/commit/2bc9035))
+
+
+
+
+
+# [8.0.0](https://github.com/pie-framework/pie-lib/compare/@pie-lib/config-ui@7.11.34...@pie-lib/config-ui@8.0.0) (2019-04-12)
+
+
+### Features
+
+* New component Settings panel for rendering the settings ui. ([7ce4741](https://github.com/pie-framework/pie-lib/commit/7ce4741))
+* new config layout component ([452e8a9](https://github.com/pie-framework/pie-lib/commit/452e8a9))
+
+
+### BREAKING CHANGES
+
+* component is accessible as `import {layout} from
+'@pie-lib/config-ui'; const {ConfigLayout}=settings;`
+
+
+
+
+
+## [7.11.34](https://github.com/pie-framework/pie-lib/compare/@pie-lib/config-ui@7.11.33...@pie-lib/config-ui@7.11.34) (2019-04-11)
+
+**Note:** Version bump only for package @pie-lib/config-ui
+
+
+
+
+
 ## [7.11.33](https://github.com/pie-framework/pie-lib/compare/@pie-lib/config-ui@7.11.32...@pie-lib/config-ui@7.11.33) (2019-04-10)
 
 **Note:** Version bump only for package @pie-lib/config-ui