+++ packages/math-toolbar/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.3](https://github.com/pie-framework/pie-lib/compare/@pie-lib/math-toolbar@0.11.2...@pie-lib/math-toolbar@0.11.3) (2019-04-12)
+
+**Note:** Version bump only for package @pie-lib/math-toolbar
+
+
+
+
+
+## [0.11.2](https://github.com/pie-framework/pie-lib/compare/@pie-lib/math-toolbar@0.11.1...@pie-lib/math-toolbar@0.11.2) (2019-04-12)
+
+**Note:** Version bump only for package @pie-lib/math-toolbar
+
+
+
+
+
+## [0.11.1](https://github.com/pie-framework/pie-lib/compare/@pie-lib/math-toolbar@0.11.0...@pie-lib/math-toolbar@0.11.1) (2019-04-11)
+
+**Note:** Version bump only for package @pie-lib/math-toolbar
+
+
+
+
+
 # [0.11.0](https://github.com/pie-framework/pie-lib/compare/@pie-lib/math-toolbar@0.10.2...@pie-lib/math-toolbar@0.11.0) (2019-04-10)